pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fsspec>=2026.2.0",
    "huggingface-hub>=1.5.0",
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.5",
    "matplotlib>=3.10.8",
    "neo4j>=6.1.0",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.47",
]
